Extracting MonteCarlo Tree Search out of gamer
