--- .settings/org.eclipse.ltk.core.refactoring.prefs
@@ -1,3 +1,0 @@
-#Tue Mar 23 10:27:05 PDT 2010
-eclipse.preferences.version=1
-org.eclipse.ltk.core.refactoring.enable.project.refactoring.history=false
